backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.2.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
bcrypt==4.1.1
PyJWT==2.8.0
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
pytest==7.4.3
pytest-django==4.7.0

